transcrição para classe/component

+++ backend/src/controllers/UpdateUserController.js
@@ -5,30 +5,31 @@ const bcrypt = require('bcrypt');
 const path = require('path');
 const fs = require('fs');
 const {promisify} = require('util');
+require('dotenv').config();
 
 module.exports = {
 
     async index(req, res, next){
 
-        //await Auth.index(req,res,next);
+        await Auth.index(req,res,next);
 
 
         const {user} = req.body
         var query = {'email': user.email};  
 
-        const {oldFile} = user.file;
+        if (user.file){
+            const {oldFile} = user.file;
         
-        if(oldFile !== null && oldFile !== '' && oldFile !== undefined){
-            const fileName = oldFile.replace('http://localhost:3333/files/', '')
-            console.log(fileName);
-            promisify(fs.unlink)(path.resolve(__dirname, '..', '..', 'uploads', fileName));
+            if(oldFile !== null && oldFile !== '' && oldFile !== undefined){
+                const fileName = oldFile.replace(process.env.BASE_URL + 'files/', '');
+                promisify(fs.unlink)(path.resolve(__dirname, '..', '..', 'uploads', fileName));
+            }
         }
         
         var updateUser = await User.findOneAndUpdate(query, user, {returnNewDocument: true}, function(error, user) {
             if (error)  {
                 return res.json({error});
             } else {
-                // follow up action
                 update = "Usuário Atualizado com sucesso !";
                 return res.json({updateUser: user, message: update});
             }
@@ -49,11 +50,11 @@ module.exports = {
         user = await User.findOne({email: userReq.email}).select('+password');
 
         if (!user)  {
-            return res.status(400).json({error: "Senha inválido"});
+            return res.status(400).send({error: "Senha inválida"});
         }
 
-        if(!await bcrypt.compare(userReq.senhaAntiga, user.password)){
-            return res.status(400).json({error: "Senha inválido"});
+        if(!await bcrypt.compare(userReq.senhaAntiga, user.password).catch((err) => {})){
+            return res.status(400).send({error: "Senha inválida"});
         }
 
         user.password =  await bcrypt.hash(userReq.senhaNova, 10);
